[PersistentCollections] Move BitmapSmokeTests into its own file
